fix/update consult
